Adds helper methods and an Excel Reader for menu
Helper includes an InlineKeyboard method to build a InlineKeyboardMarkup
from a list of items. Button names and callback_data are set to
be the same.

ExcelBuilder contains a menu class, allows the user to encapsulate the
list of food using an object.
Method list retrieves a list of food available.
Method find_item returns the row which contains the item as a list. The
format of the list is (orderID, itemname, option 1, option 2....)
Method item_stages returns the list of options for the item chosen.
Format of the list is ("option 1A,option 1B", "option 2A,option 2B"...).

__init__ has initialized as a blank file.