Improves data rendering using Pandas dataframe
Data has a menu object which builds a dataframe using Pandas from
an excel file. The Menu class has 3 methods:
1. items is a method which takes in the restaurant(str)
and object id(str) to return the name of the object and it's options.
options(int) is a keyword argument which allows the user to select which
data they want: 0 - Item, 1 - Option 1, 2 - Option 2.
2. price is a method which takes in the restaurant(str)
and object id(str) to return the price of the object and it's options.
3. listing takes in a restaurant(str), data type(int) and
option type(int). For data type, there are 3 types: 0 - Object ID,
1 - Object's name, 2- Object's price. For option type there are 3 types:
0 - Item, 1 - Option 1, 2- Option 2.

ExcelBuilder is replaced with Data while helpers has been removed
as it is useful.

Adds a Menu.xlsx for testing purposes. Data automaticallly reads
Menu.xlsx to build the dataframe. To import dataframe just import
"menu".