Moves handlers outside of class and tags prices
addOrder, removeOrder and viewOrder and their respective helper
functions were moved outside of the class. In order to navigate and know
which order is associate with which chat group, handlers use bot_data
to store the chat_id for each incoming user.

Bot now only accepts commands called from within the group instead
of via direct message with the bot. It gives an error message if user
attempts to chat with the bot directly. This is to allow the bot to know
which chat the user is ordering from.

Prices has been included in the menu options (like a normal menu).The
user will also be able to see their total price when calling viewOrder.

Adds removeOrder method to order class to removeorders from food data.

Food data saves the orderID instead of the item name, this is to aid in
lookup for the price better.