Changes startjio to letsmakan & adds new questions
Startjio and CloseJio have been changed to LetsMakan and EndMakan
respectively.

Instead of using a master_list_of_jios to store the dictionary of
orders, bot_data is used instead. Chat_id as the key and order object
as the value.

Adds two more questions: Phone number and address. These information
will be intialised in the order object.

AddsPreOrderinStageHandlers method to build and add preordering stage
handlers to dispatcher.