Initialises bot data and removes persistence
Bot data contains a dictionary of store details. The keys are
the store names and values are dictionaries of its cache.
dict['Store Open'] has a boolean value to indicate if the store is open.
dict['orders'] has a queue which represents a queue of orders via
python's native queue module

Removes Pickle Persistence as it is causing errors such as pickel dump.