Adds tuple inputs in Data's item and cost methods
ID in item and cost methods can either be a int or tuple. For int
inputs, item and cost returns the corresponding names and costs for the
ID, with default option = 0. Item and cost can also read tuples and
combines the respective names and cost.

e.g ID = ("A1", 0, 2) combines "A1", option_1 = 0 and option _2 = 2 to
give either the concatenated name or total sum of the item's price.
This implmentation allows orderingstage to easily query for names and
prices when there are multiple options. In the event there are
few options than 3, the missing option should be
initialised to None. e.g ("A1", 0, None)

Adds Test and simulation methods to test for errors in implementation.