Adds secondary options to Addorder method
If there are secondary options present for the chosen item in Addorder,
the bot will edit the existing message and change the reply_markup
to show new secondary options to choose from. The current number of
secondary options is 2 for simplicity sake.

Addorder conversation handler is builds with options. Options is called
twice or less depending on the number of options. When there are no more
options, addorder_helper is called to combine all the option names and
prices.

Dictionaries of food in order objects uses tuples as keys instead of
strings to accomadate for the options. To updateList and removefood,
the ID should be of type tuple.

Fixes removeOrder. Bot prompts user to that he has not added anything
if removeOrder is called when the user has not ordered yet.