Adds accepted member and printOrder method.
Class Order has 'accepted' member to track the status of the order,
if its been processed, accepted or rejected. accepted = None if it is
not processed yet, True if its been processed and accepted by the Store
and False if its been processed and rejected by the Store.

printOrder returns a the list of orders as a string.

User's are prompted with a /cancel command should they want to cancel
the current action (e.g addOrder and rejectOrder). The cancel method
exits the conversation and clear the user cache before ending it. If
there are any inline keyboard, it clears them as well.