Changes convo structure.
Phone and address are requested by the bot after the user calls
EndMakan instead of when it is called during LetsMakan.

LetsMakan checks if restaurants are available before asking user to
choose a restaurant. In the event that there are not restaurants
available, the bot notifies the user. This code has been commented out
for testing purposes.