Adds poll feature and checks available stores
Adds a poll command to allow users to vote for which restaurant they
want to choose. receive_poll_answer handles poll answers and updates
the chat group which member selected which restaurant.

The poll stops when all members have voted. In the event that there are
other bots within the group, the poll will not close.

Before the bot queries users for which restaurant to choose, be it
using poll or letsmakan, the bot first checks for avaialble stores.
In the event that there stores are closed, the bot notifies the
user that there are not restaurants opened. If there is only one
store open, poll does not send a poll to the chat group, but notifies
user to call letsmakan to start the supper session.