Adds check feature for numbers and addresses
Only valid numbers are accepted when bot queries for numbers, otherwise
the bot will ask the user to enter a which starts with (9 or 8 or 6)
with length of 8.

When asking for addresses, the bot gives a menu of the list of halls to
select from, rather than asking the user to type it out.